+++ src/app/modules/index.ts
@@ -1,10 +1,12 @@
 
 export * from './settings/settings.module';
 export * from './gpxload/gpxload.module'
+export * from './geojson/geojson.module'
 export * from './alarms/alarms.module'
 export * from './skresources/resources.module'
 export * from './map/map.module'
 export * from './utils/utils.module'
 export * from './experiments/experiments.module'
+export * from './course/course.module'
 
 export * from './skstream/skstream.module'